+++ README.md
@@ -1,33 +1,32 @@
-# movepy
+# refac
 
-A tool for moving and refactoring python code. It supports moving files or entire directories (movepy file). Also supports moving specific symbols inside a file (movepy symbol) or specific imports (movepy import).
+A tool for moving and refactoring python code. It supports moving files or entire directories (refac file). Also supports moving specific symbols inside a file (refac symbol) or specific imports (refac import).
 
-[![Checks](https://github.com/benchling/movepy/actions/workflows/checks.yaml/badge.svg?branch=main&event=push)](https://github.com/benchling/movepy/actions/workflows/checks.yaml)
+[![Checks](https://github.com/benchling/refac/actions/workflows/checks.yaml/badge.svg?branch=main&event=push)](https://github.com/benchling/refac/actions/workflows/checks.yaml)
 
 ## Quickstart
 
 ```bash
-pip install movepy
-movepy file /path/to/src.py /path/to/dst.py
+pip install refac
+refac file /path/to/src.py /path/to/dst.py
 ```
 
 and add the following to your `.libcst.codemod.yaml` at the root of your project:
 
 ```yaml
 modules:
-- 'movepy.visitors'
+  - "refac.visitors"
 ```
-
 
 ```bash
 usage:
-    movepy [file|symbol|import] <src> <dst>
+    refac [file|symbol|import] <src> <dst>
 
   examples:
-    movepy file /path/to/src.py /path/to/dst.py
-    movepy symbol path.to.SrcClass path.to.DstClass
-    movepy symbol path.to.src_func1,path.to.src_func2 path.to.dst_func1,path.to.dst_func2
-    movepy import path.to.src_import path.to.dst_import
+    refac file /path/to/src.py /path/to/dst.py
+    refac symbol path.to.SrcClass path.to.DstClass
+    refac symbol path.to.src_func1,path.to.src_func2 path.to.dst_func1,path.to.dst_func2
+    refac import path.to.src_import path.to.dst_import
 ```
 
 ## Contributing
@@ -35,7 +34,6 @@ usage:
 Contributions are welcomed and appreciated. Check out ARCHITECTURE.md for an overview of the codebase.
 
 To run tests:
-
 
 ```bash
 # Must be using Python >3.9
