--- src/movepy/__main__.py
@@ -1,3 +1,0 @@
-from movepy import main
-
-main()
