+++ src/refac/__main__.py
@@ -1,0 +1,3 @@
+from refac import main
+
+main()
