+++ src/refac/move_symbol.py
@@ -10,10 +10,10 @@ from libcst.codemod._context import CodemodContext
 from libcst.metadata.full_repo_manager import FullRepoManager
 from libcst.metadata.name_provider import FullyQualifiedNameProvider
 
-from movepy.replace_str import find_and_replace
-from movepy.utils import ROOT_DIR, shell, to_file
-from movepy.visitors.add_symbols import AddSymbolsVisitor
-from movepy.visitors.remove_symbols import RemoveSymbolsVisitor
+from refac.replace_str import find_and_replace
+from refac.utils import ROOT_DIR, shell, to_file
+from refac.visitors.add_symbols import AddSymbolsVisitor
+from refac.visitors.remove_symbols import RemoveSymbolsVisitor
 
 
 def validate(old_full_symbols: List[str], new_full_symbols: List[str]) -> None:
