+++ src/refac/visitors/remove_symbols.py
